Fix integration tests
Two issues were causing the test suite to fail intermitently; firstly the
ServiceUtils class was keeping a cache of a TransactionManager.  If
the TransactionManagerLocatorTest ran first, the TransactionManager was
cached in ServiceUtils resulting in lookup never being called on the
WildFlyTransactionManagerLocator in any subsequent test.

Secondly there existed 2 @Before methods declared on the
WildFlyRecoveryRegistry tests.  One was inherited from
ActiveMQRAClusteredTest.  Depeding on the order of the @before methods
(not enforced by JUnit) the tests may pass or fail (as the JNDI
entry would or for the TM would not exist before the lookuo).